surveillance/requirements.txt
arrow
asgiref
branca
certifi
charset-normalizer
Django
folium
idna
imutils
Jinja2
MarkupSafe
numpy
opencv-contrib-python
pandas
Pillow
playsound
psycopg2-binary
# pycairopip
PyGObject
python-dateutil
pytz
requests
six
sqlparse
times
urllib3
absl-py
astunparse
cachetools
chardet
click
Cython
face-recognition-models
Flask
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
importlib-metadata
itsdangerous
joblib
Keras-Preprocessing
Markdown
oauthlib
opt-einsum
protobuf
pyasn1
pyasn1-modules
PySide2
pytesseract
requests-oauthlib
rsa
scikit-learn
scipy
shiboken2
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor
tesseract
Werkzeug
wrapt
zipp
